This entire pull request is never meant to be merged. I had the opportunity to use the code base here as a proof of concept for an F# to C# conversion for a company hackathon. Everything you see here was done in one day.
The conversion of the common library to F# was pretty straight forward. There was the slight annoyance of adding the "With" methods, since C# had no direct way to easily mutate an F# record.

The next annoyance was creating Tasks in F# for C# to consume. That was mostly just figuring out the flow.

I managed to get the StudentCodeDataStore converted. This allowed me to drop it into the TicketController. This should be the most interesting part to look at. Just how little code changed still floors me.

Hope you find this interesting.